requirements.txt
aiohttp
DrissionPage
pyautogui
pillow
google-generativeai
python-dotenv
fastapi 
uvicorn
pyaudio
google-genai